feat(transaction): add clientTransaction for typed clients without verb mirrors
clientTransaction binds useTransaction to a server defineTransaction via
import type, so the browser only declares verb names/kinds instead of
duplicating inert callbacks. Consolidates the blog-studio example into a
shared lib module and updates docs, skill, and e2e coverage.
