+++ testing/e2e/src/routes/api.transaction.ts
@@ -27,6 +27,69 @@ async function collectText(stream: AsyncIterable<unknown>): Promise<string> {
   return text
 }
 
+type TextOptions = ReturnType<typeof createTextAdapter>
+
+/** Build the e2e transaction for a resolved provider/test routing context. */
+export function createE2eTransaction(textOptions: () => TextOptions) {
+  // One-shot verb: schema-validated input in, result out. Runs a
+  // single deterministic chat completion against aimock (the fixture
+  // keys on the exact `[transaction] banner: …` user message).
+  const banner = verb({
+    input: z.object({ prompt: z.string() }),
+    execute: async (req) => {
+      const text = await collectText(
+        chat({
+          ...textOptions(),
+          messages: [
+            {
+              role: 'user',
+              content: `[transaction] banner: ${req.input.prompt}`,
+            },
+          ],
+          agentLoopStrategy: maxIterations(1),
+          threadId: req.threadId,
+          runId: req.runId,
+        }),
+      )
+      return { prompt: req.input.prompt, text }
+    },
+  })
+
+  return defineTransaction({
+    // Conversational verb: full chat surface on the client.
+    primaryChat: chatVerb((req) =>
+      chat({
+        ...textOptions(),
+        messages: req.messages,
+        agentLoopStrategy: maxIterations(5),
+        threadId: req.threadId,
+        runId: req.runId,
+      }),
+    ),
+    banner,
+    // Composing verb: runs the sibling `banner` verb twice via
+    // `ctx.call`, so the client observes two tagged sub-runs
+    // (transaction:sub-run:* CUSTOM events) inside one SSE response.
+    bannerPair: verb({
+      input: z.object({ topic: z.string() }),
+      execute: async ({ input }, ctx) => {
+        // Sequential so sub-run start order (hero=0, thumb=1) is
+        // deterministic for the spec's per-index assertions.
+        const hero = await ctx.call(banner, {
+          prompt: `hero ${input.topic}`,
+        })
+        const thumb = await ctx.call(banner, {
+          prompt: `thumb ${input.topic}`,
+        })
+        return { hero, thumb }
+      },
+    }),
+  })
+}
+
+/** Server-side transaction shape — import with `import type` on the client. */
+export type E2eTransaction = ReturnType<typeof createE2eTransaction>
+
 export const Route = createFileRoute('/api/transaction')({
   server: {
     handlers: {
@@ -54,60 +117,7 @@ export const Route = createFileRoute('/api/transaction')({
             'transaction',
           )
 
-        // One-shot verb: schema-validated input in, result out. Runs a
-        // single deterministic chat completion against aimock (the fixture
-        // keys on the exact `[transaction] banner: …` user message).
-        const banner = verb({
-          input: z.object({ prompt: z.string() }),
-          execute: async (req) => {
-            const text = await collectText(
-              chat({
-                ...textOptions(),
-                messages: [
-                  {
-                    role: 'user',
-                    content: `[transaction] banner: ${req.input.prompt}`,
-                  },
-                ],
-                agentLoopStrategy: maxIterations(1),
-                threadId: req.threadId,
-                runId: req.runId,
-              }),
-            )
-            return { prompt: req.input.prompt, text }
-          },
-        })
-
-        const transaction = defineTransaction({
-          // Conversational verb: full chat surface on the client.
-          primaryChat: chatVerb((req) =>
-            chat({
-              ...textOptions(),
-              messages: req.messages,
-              agentLoopStrategy: maxIterations(5),
-              threadId: req.threadId,
-              runId: req.runId,
-            }),
-          ),
-          banner,
-          // Composing verb: runs the sibling `banner` verb twice via
-          // `ctx.call`, so the client observes two tagged sub-runs
-          // (transaction:sub-run:* CUSTOM events) inside one SSE response.
-          bannerPair: verb({
-            input: z.object({ topic: z.string() }),
-            execute: async ({ input }, ctx) => {
-              // Sequential so sub-run start order (hero=0, thumb=1) is
-              // deterministic for the spec's per-index assertions.
-              const hero = await ctx.call(banner, {
-                prompt: `hero ${input.topic}`,
-              })
-              const thumb = await ctx.call(banner, {
-                prompt: `thumb ${input.topic}`,
-              })
-              return { hero, thumb }
-            },
-          }),
-        })
+        const transaction = createE2eTransaction(textOptions)
 
         return transaction.handler(request)
       },
